infrastructure/aws/requirements.txt
aws-cdk-lib==2.173.2
constructs>=10.0.0
